[FIX] Refactored navigation manager and actuator
