+++ py/examples/waypoint_navigation/visualization/robot_positions.json
@@ -1,20 +1,8 @@
 [
   {
     "segment_name": "waypoint_0_to_1",
-    "x": 11113.152824272265,
-    "y": 8425.13207077352,
-    "heading": -0.19392243807899184
-  },
-  {
-    "segment_name": "waypoint_1_to_2",
-    "x": 11116.23973451558,
-    "y": 8423.630269260657,
-    "heading": -0.0533833415556361
-  },
-  {
-    "segment_name": "waypoint_2_to_3",
-    "x": 11125.016153535682,
-    "y": 8422.419229156372,
-    "heading": -0.5327257240233269
+    "x": 11113.074912436936,
+    "y": 8424.545345480847,
+    "heading": -0.13299115836741784
   }
 ]
